Fixed a real bug in HitCheater module and a potential one in HoughBaseAlg and the same HitCheater. Both introduced with new data products.

+++ RecoAlg/HoughBaseAlg.cxx
@@ -1777,7 +1777,8 @@ size_t cluster::HoughBaseAlg::FastTransform(const std::vector<art::Ptr<recob::Cl
       clusterHits.clear();    
       if(lastHits.size() < 5) continue;
       
-      lar::util::StatCollector<float> integralQ, summedQ;
+      // reduce rounding errors by using double (RMS is very sensitive to them)
+      lar::util::StatCollector<double> integralQ, summedQ;
       
       for(size_t i = 0; i < lastHits.size(); ++i) {
         clusterHits.push_back(hit.at(hitTemp.at(lastHits.at(i))));
